refactor(typescript): types for navigation (navbar list, modal, dots)
