+++ src/constants/navList.ts
@@ -1,0 +1,17 @@
+enum NavList {
+    HOMEPAGE = "accueil",
+    ABOUT = "a-propos",
+    SKILLS = "competences",
+    PORTFOLIO = "portfolio",
+    HOBBIES = "hobbies",
+    CONTACT = "contact",
+}
+
+export const navList: string[] = [
+    NavList.HOMEPAGE,
+    NavList.ABOUT,
+    NavList.SKILLS,
+    NavList.PORTFOLIO,
+    NavList.HOBBIES,
+    NavList.CONTACT
+]
